Add auto-starring functionality and exit on complete
Add a Star() method to the GitHubStargazer, and allow the user to set a
token when initializing the stargazer. Add a Stop() method to exit the
Gaze() loop.

Modify the on-threshold-crossed function to attempt to star the watched
repository, and then exit.

Idea: don't exit, and add a hook for threshold being crossed in the
downward direction, and unstar the repo if this happens.